fix(UX): change column name to Variant SKUs in product table
In the "Products in Shopify" table,  changed the column name from "SKUs" to "Variant SKUs" to avoid naming confusion between ID (parent) and SKU (variants). 
Shopify uses product_id, variant_ids and sku to define the hierarchy of parent and child products while in ERPNext we just have item_code as the UID for parent item as well as the variant item.